feat: this module is now named `ipld-block`
This module was previously called `ipfs-block`. It is now maintained within IPLD.

BREAKING CHANGE: The symbol for comparing blocks changed

This means that if you are a user of this module, updating to this version
is also a breaking change for your module. Don't forgot to make the corresponding
version bump on your next release.